+++ lib/src/internal/db/schema/channel/c_group_channel.g.dart
@@ -320,7 +320,7 @@ final CGroupChannelSchema = CollectionSchema(
   getId: _cGroupChannelGetId,
   getLinks: _cGroupChannelGetLinks,
   attach: _cGroupChannelAttach,
-  version: '3.1.0+1',
+  version: '3.3.0-dev.2',
 );
 
 int _cGroupChannelEstimateSize(