+++ lib/src/internal/db/schema/chunk/c_base_chunk.dart
@@ -1,6 +1,6 @@
 // Copyright (c) 2023 Sendbird, Inc. All rights reserved.
 
-import 'package:isar/isar.dart';
+import 'package:isar_community/isar.dart';
 import 'package:sendbird_chat_sdk/src/internal/db/schema/chunk/base_chunk.dart';
 import 'package:sendbird_chat_sdk/src/internal/db/schema/chunk/chunk.dart';
 import 'package:sendbird_chat_sdk/src/internal/main/chat/chat.dart';