+++ lib/src/internal/db/schema/message/c_admin_message.g.dart
@@ -261,7 +261,7 @@ final CAdminMessageSchema = CollectionSchema(
   getId: _cAdminMessageGetId,
   getLinks: _cAdminMessageGetLinks,
   attach: _cAdminMessageAttach,
-  version: '3.1.0+1',
+  version: '3.3.0-dev.2',
 );
 
 int _cAdminMessageEstimateSize(