+++ lib/src/internal/db/schema/message/c_file_message.g.dart
@@ -302,7 +302,7 @@ final CFileMessageSchema = CollectionSchema(
   getId: _cFileMessageGetId,
   getLinks: _cFileMessageGetLinks,
   attach: _cFileMessageAttach,
-  version: '3.1.0+1',
+  version: '3.3.0-dev.2',
 );
 
 int _cFileMessageEstimateSize(