+++ lib/src/internal/db/schema/message/c_user_message.g.dart
@@ -280,7 +280,7 @@ final CUserMessageSchema = CollectionSchema(
   getId: _cUserMessageGetId,
   getLinks: _cUserMessageGetLinks,
   attach: _cUserMessageAttach,
-  version: '3.1.0+1',
+  version: '3.3.0-dev.2',
 );
 
 int _cUserMessageEstimateSize(