+++ lib/src/internal/main/model/log_publish_realtime_config.g.dart
@@ -9,9 +9,10 @@ part of 'log_publish_realtime_config.dart';
 LogPublishRealtimeConfig _$LogPublishRealtimeConfigFromJson(
         Map<String, dynamic> json) =>
     LogPublishRealtimeConfig(
-      minStatCount: json['min_stat_count'] as int?,
-      minInterval: json['min_interval'] as int?,
-      maxStatCountPerRequest: json['max_stat_count_per_request'] as int?,
-      lowerThreshold: json['lower_threshold'] as int?,
-      requestDelayRange: json['request_delay_range'] as int?,
+      minStatCount: (json['min_stat_count'] as num?)?.toInt(),
+      minInterval: (json['min_interval'] as num?)?.toInt(),
+      maxStatCountPerRequest:
+          (json['max_stat_count_per_request'] as num?)?.toInt(),
+      lowerThreshold: (json['lower_threshold'] as num?)?.toInt(),
+      requestDelayRange: (json['request_delay_range'] as num?)?.toInt(),
     );