+++ lib/src/internal/network/websocket/event/login_event.g.dart
@@ -10,19 +10,19 @@ LoginEvent _$LoginEventFromJson(Map<String, dynamic> json) => LoginEvent(
       eKey: json['ekey'] as String?,
       newKey: json['new_key'] as String?,
       key: json['key'] as String?,
-      loginTimestamp: json['login_ts'] as int,
+      loginTimestamp: (json['login_ts'] as num).toInt(),
       reconnectConfiguration: ReconnectConfiguration.fromJson(
           json['reconnect'] as Map<String, dynamic>),
-      pingInterval: json['ping_interval'] as int? ?? 15,
-      watchdogInterval: json['pong_timeout'] as int,
+      pingInterval: (json['ping_interval'] as num?)?.toInt() ?? 15,
+      watchdogInterval: (json['pong_timeout'] as num).toInt(),
       maxUnreadCountOnSuperGroup:
-          json['max_unread_count_on_super_group'] as int?,
+          (json['max_unread_count_on_super_group'] as num?)?.toInt(),
       applicationAttributes: (json['application_attributes'] as List<dynamic>?)
               ?.map((e) => e as String)
               .toList() ??
           const [],
       deviceTokenLastDeletedAt:
-          json['device_token_last_deleted_at'] as int? ?? 0,
+          (json['device_token_last_deleted_at'] as num?)?.toInt() ?? 0,
       configSyncNeeded: json['config_sync_needed'] as bool?,
       logPublishConfig: json['log_publish_config'] == null
           ? null